Fix search bar error and refactor api file
