+++ admin_app/src/utils/api.ts
@@ -11,7 +11,7 @@ const api = axios.create({
   },
 });
 
-import { AxiosResponse, AxiosError } from 'axios';
+import { AxiosResponse, AxiosError } from "axios";
 
 api.interceptors.response.use(
   (response: AxiosResponse) => response,
@@ -27,12 +27,6 @@ api.interceptors.response.use(
   },
 );
 
-interface ContentBody {
-  content_title: string;
-  content_text: string;
-  content_metadata: Record<string, unknown>;
-}
-
 const getUser = async (token: string) => {
   try {
     const response = await api.get("/user/", {
@@ -43,160 +37,6 @@ const getUser = async (token: string) => {
     return response.data;
   } catch (error) {
     throw new Error("Error fetching user info");
-  }
-};
-
-const getContentList = async ({
-  token,
-  skip = 0,
-  limit = 200,
-}: {
-  token: string;
-  skip?: number;
-  limit?: number;
-}) => {
-  try {
-    const response = await api.get(`/content/?skip=${skip}&limit=${limit}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error fetching content list");
-  }
-};
-
-const getContent = async (content_id: number, token: string) => {
-  try {
-    const response = await api.get(`/content/${content_id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error fetching content");
-  }
-};
-
-const archiveContent = async (content_id: number, token: string) => {
-  try {
-    const response = await api.patch(
-      `/content/${content_id}`,
-      {},
-      {
-        headers: { Authorization: `Bearer ${token}` },
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw new Error("Error archiving content");
-  }
-};
-
-const deleteContent = async (content_id: number, token: string) => {
-  try {
-    const response = await api.delete(`/content/${content_id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error deleting content");
-  }
-};
-
-const editContent = async (content_id: number, content: ContentBody, token: string) => {
-  try {
-    const response = await api.put(`/content/${content_id}`, content, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error editing content");
-  }
-};
-
-const createContent = async (content: ContentBody, token: string) => {
-  try {
-    const response = await api.post("/content/", content, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error creating content");
-  }
-};
-
-const bulkUploadContents = async (file: File, token: string) => {
-  const formData = new FormData();
-  formData.append("file", file);
-
-  try {
-    const response = await api.post("/content/csv-upload", formData, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-        "Content-Type": "multipart/form-data",
-      },
-    });
-
-    return { status: response.status, data: response.data };
-  } catch (error: any) {
-    if (error.response) {
-      const errorResponse = error.response.data;
-      const detail = errorResponse.detail || "An unknown error occurred";
-      return { status: error.response.status, detail: detail };
-    } else {
-      throw new Error("An unexpected error occurred during the bulk upload.");
-    }
-  }
-};
-
-const getUrgencyRuleList = async (token: string) => {
-  try {
-    const response = await api.get("/urgency-rules/", {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error fetching urgency rule list");
-  }
-};
-
-const addUrgencyRule = async (rule_text: string, token: string) => {
-  try {
-    const response = await api.post(
-      "/urgency-rules/",
-      { urgency_rule_text: rule_text },
-      {
-        headers: { Authorization: `Bearer ${token}` },
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw new Error("Error adding urgency rule");
-  }
-};
-
-const updateUrgencyRule = async (rule_id: number, rule_text: string, token: string) => {
-  try {
-    const response = await api.put(
-      `/urgency-rules/${rule_id}`,
-      { urgency_rule_text: rule_text },
-      {
-        headers: { Authorization: `Bearer ${token}` },
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw new Error("Error updating urgency rule");
-  }
-};
-
-const deleteUrgencyRule = async (rule_id: number, token: string) => {
-  try {
-    const response = await api.delete(`/urgency-rules/${rule_id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error deleting urgency rule");
   }
 };
 
@@ -231,13 +71,11 @@ const getGoogleLoginToken = async (idToken: {
     throw new Error("Error fetching Google login token");
   }
 };
-
-// Perform search
 const getSearch = async (
   question: string,
   generate_llm_response: boolean,
   token: string,
-) => {
+): Promise<{ status: number; data?: any; error?: any }> => {
   try {
     const response = await api.post(
       "/search",
@@ -249,13 +87,19 @@ const getSearch = async (
         headers: { Authorization: `Bearer ${token}` },
       },
     );
+
     return { status: response.status, ...response.data };
-  } catch (error) {
-    throw new Error("Error performing search");
+  } catch (err) {
+    const error = err as AxiosError;
+    if (error.response) {
+      return { status: error.response.status, error: error.response.data };
+    } else {
+      console.error("Error performing search", error.message);
+      throw new Error(`Error performing search: ${error.message}`);
+    }
   }
 };
 
-// Post response feedback
 const postResponseFeedback = async (
   query_id: number,
   feedback_sentiment: string,
@@ -280,19 +124,6 @@ const postResponseFeedback = async (
   }
 };
 
-// Get question stats
-const getQuestionStats = async (token: string) => {
-  try {
-    const response = await api.get("/dashboard/question_stats", {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error fetching question stats");
-  }
-};
-
-// Get urgency detection
 const getUrgencyDetection = async (search: string, token: string) => {
   try {
     const response = await api.post(
@@ -308,66 +139,12 @@ const getUrgencyDetection = async (search: string, token: string) => {
   }
 };
 
-// Create new tag
-const createTag = async (tag: string, token: string) => {
-  try {
-    const response = await api.post(
-      "/tag/",
-      { tag_name: tag },
-      {
-        headers: { Authorization: `Bearer ${token}` },
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw new Error("Error creating tag");
-  }
-};
-
-// Get tag list
-const getTagList = async (token: string) => {
-  try {
-    const response = await api.get("/tag/", {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error fetching tag list");
-  }
-};
-
-// Delete tag
-const deleteTag = async (tag_id: number, token: string) => {
-  try {
-    const response = await api.delete(`/tag/${tag_id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return response.data;
-  } catch (error) {
-    throw new Error("Error deleting tag");
-  }
-};
 export const apiCalls = {
   getUser,
-  getContentList,
-  getContent,
-  archiveContent,
-  deleteContent,
-  editContent,
-  createContent,
-  bulkUploadContents,
-  getUrgencyRuleList,
-  addUrgencyRule,
-  updateUrgencyRule,
-  deleteUrgencyRule,
   getLoginToken,
   getGoogleLoginToken,
   getSearch,
   postResponseFeedback,
-  getQuestionStats,
   getUrgencyDetection,
-  createTag,
-  getTagList,
-  deleteTag,
 };
 export default api;
